+++ worldeditadditions_commands/commands/nodes/init.lua
@@ -1,0 +1,11 @@
+-- ███    ██  ██████  ██████  ███████ ███████
+-- ████   ██ ██    ██ ██   ██ ██      ██     
+-- ██ ██  ██ ██    ██ ██   ██ █████   ███████
+-- ██  ██ ██ ██    ██ ██   ██ ██           ██
+-- ██   ████  ██████  ██████  ███████ ███████
+
+-- Commands that operate on information about nodes, and NOT on the world itself..
+
+local we_cmdpath = worldeditadditions_commands.modpath .. "/commands/nodes/"
+
+dofile(we_cmdpath .. "ngroups.lua")